Remove old quickstart, update test name
